+++ src/main/java/com/zaia08/hub_estudos/controller/AlterCursoDTO.java
@@ -2,4 +2,4 @@ package com.zaia08.hub_estudos.controller;
 
 import java.time.LocalDate;
 
-public record AlterCursoDTO(String nome_curso, Integer horas_desejadas, Integer horas_atuais, LocalDate meta_de_conclusao, Boolean resumo_semanal) {}
+public record AlterCursoDTO(String nome_curso, Integer horas_desejadas, Float horas_atuais, LocalDate meta_de_conclusao, Boolean resumo_semanal) {}
