+++ src/main/java/com/zaia08/hub_estudos/controller/CreateCursoDTO.java
@@ -2,4 +2,4 @@ package com.zaia08.hub_estudos.controller;
 
 import java.time.LocalDate;
 
-public record CreateCursoDTO(String nome_curso, Integer horas_desejadas, Integer horas_atuais, LocalDate meta_de_conclusao, Boolean resumo_semanal, String prioridade) {}
+public record CreateCursoDTO(String nome_curso, Integer horas_desejadas, Float horas_atuais, LocalDate meta_de_conclusao, Boolean resumo_semanal, String prioridade) {}
